Update NumberBytesConvertUtil.java
Fix a bug of convert byte array to long

+++ src/main/java/com/macrohuang/fileq/util/NumberBytesConvertUtil.java
@@ -65,7 +65,7 @@ public class NumberBytesConvertUtil {
 		}
 		long result = 0;
 		for (int i = 0; i < bytes.length; i++) {
-			result |= bytes[i] << ((bytes.length - 1 - i) * 8) & LONG_BASE[LONG_BASE.length - bytes.length + i];
+			result |= (long)bytes[i] << ((bytes.length - 1 - i) * 8) & LONG_BASE[LONG_BASE.length - bytes.length + i];
 		}
 		return result;
 	}
